requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-ipware==4.0.2
django-taggit==3.0.0
future==0.18.2
geocoder==1.38.1
idna==3.3
Pillow==9.2.0
ratelim==0.1.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
